bam_cram_validate_compare.sh - Add more strict filters to verify file found for the bam and cram validation output or it will fail.  Also removal of BAM dir I chose to use the "BAM_MAIN_DIR/BAM"... but that doesn't catch the BAM_BAI or other alternatives there could be.  Plain ol' $BAM_DIR should have been used from the get go.
bam_to_cram_remove_tags.sh (and rnd) - changed the file size calculation to automatically put into gigs for ease.

md5_check.sh - Here any file path that has a space involved, be it a directory or the file itself.  This temporary change in the internal field separator (IFS) should fix this (worked in testing).  IFS is then changed back at the end of the script.  Also, use the -type f to find only files.... Ran into an issue with a folder ending in vcf which didn't play nicely with my original setup.

zip_file.sh - removed unnecessary comment lines.

compressor_for_rnd.sh - moved the creation of the cram_conversion_validation.list and implemented the case switch to alternate between RND and RODEN... Need to find a universal flag to say if its RND use the binning and if not don't.  This compressor script might be a universal script and not just RnD.  Also, added the -type f flag to find only files and not directories